Fixing wrong data type in invoice.qty
git-svn-id: svn://svn.code.sf.net/p/ledger-smb/code/branches/1.2@1065 4979c152-3d1c-0410-bac9-87ea11338e46

+++ sql/Pg-database.sql
@@ -101,7 +101,7 @@ CREATE TABLE invoice (
   trans_id int,
   parts_id int,
   description text,
-  qty integer,
+  qty NUMERIC,
   allocated integer,
   sellprice NUMERIC,
   fxsellprice NUMERIC,
